+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup
 VERSION = '0.0.1'
 
 description = (
-    'Include common bot functionality via custom Python packages.'
+    'Include common chatbot functionality.'
 )
 
 install_requires = ['requests']
@@ -14,10 +14,10 @@ packages = [
 ]
 
 setup(
-    name='pip-skore',
+    name='skore-chatbot-sdk',
     version=VERSION,
     description=description,
-    url='https://github.com/skore-io/pip-skore',    
+    url='https://github.com/skore-io/skore-chatbot-sdk',    
     author='skore.io',
     author_email='support@skore.io',
     license='None',
